+++ buildSrc/src/main/groovy/Config.groovy
@@ -14,8 +14,8 @@ class Config {
     static compileSdkVersion = 28
     static minSdkVersion = 14
     static targetSdkVersion = 28
-    static versionCode = 1_025_005
-    static versionName = '1.25.5'// E.g. 1.9.72 => 1,009,072
+    static versionCode = 1_025_006
+    static versionName = '1.25.6'// E.g. 1.9.72 => 1,009,072
 
     // lib version
     static kotlin_version = '1.3.10'
@@ -29,7 +29,7 @@ class Config {
 
     static depConfig = [
             plugin           : [
-                    gradle : new DepConfig("com.android.tools.build:gradle:3.3.0"),
+                    gradle : new DepConfig("com.android.tools.build:gradle:3.4.0"),
                     kotlin : new DepConfig("org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"),
                     maven  : new DepConfig("com.github.dcendents:android-maven-gradle-plugin:2.1"),// 上传到 maven
                     bintray: new DepConfig("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4"),// 上传到 bintray
@@ -37,9 +37,9 @@ class Config {
 
                     // 本地第一次上传插件新的版本需设置 useLocal = true, isApply = false
                     // 本地上传成功之后 isApply = true 即可应用插件来调试，后续版本更新无需设置 isApply = false
-                    // 发布版本的话把 useLocal = false, isApply = false，发布成功后 isApply = true 即可使用远程库版本
+                    // 发布版本的话把 useLocal = false, isApply = false，更新版本号，发布成功后 isApply = true 即可使用远程库版本
                     api    : new DepConfig(false/*是否本地调试*/, "com.blankj:api-gradle-plugin:1.0", true/*是否使用插件*/),
-                    bus    : new DepConfig(false/*是否本地调试*/, "com.blankj:bus-gradle-plugin:2.0", true/*是否使用插件*/),
+                    bus    : new DepConfig(false/*是否本地调试*/, "com.blankj:bus-gradle-plugin:2.1", true/*是否使用插件*/),
             ],
 
             api_gradle_plugin: new DepConfig(":plugin:api-gradle-plugin", false),