+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/BusUtils.java
@@ -31,7 +31,7 @@ public final class BusUtils {
     private static final Object NULL = "nULl";
     private static final String TAG  = "BusUtils";
 
-    private final Map<String, BusInfo> mTag_BusInfoMap = new HashMap<>();
+    private final Map<String, List<BusInfo>> mTag_BusInfoListMap = new HashMap<>();
 
     private final Map<String, Set<Object>>         mClassName_BusesMap          = new ConcurrentHashMap<>();
     private final Map<String, List<String>>        mClassName_TagsMap           = new HashMap<>();
@@ -50,7 +50,18 @@ public final class BusUtils {
     private void registerBus(String tag,
                              String className, String funName, String paramType, String paramName,
                              boolean sticky, String threadMode) {
-        mTag_BusInfoMap.put(tag, new BusInfo(className, funName, paramType, paramName, sticky, threadMode));
+        registerBus(tag, className, funName, paramType, paramName, sticky, threadMode, 0);
+    }
+
+    private void registerBus(String tag,
+                             String className, String funName, String paramType, String paramName,
+                             boolean sticky, String threadMode, int priority) {
+        List<BusInfo> busInfoList = mTag_BusInfoListMap.get(tag);
+        if (busInfoList == null) {
+            busInfoList = new ArrayList<>();
+            mTag_BusInfoListMap.put(tag, busInfoList);
+        }
+        busInfoList.add(new BusInfo(className, funName, paramType, paramName, sticky, threadMode, priority));
     }
 
     public static void register(final Object bus) {
@@ -87,7 +98,7 @@ public final class BusUtils {
 
     @Override
     public String toString() {
-        return "BusUtils: " + mTag_BusInfoMap;
+        return "BusUtils: " + mTag_BusInfoListMap;
     }
 
     private static BusUtils getInstance() {
@@ -112,15 +123,16 @@ public final class BusUtils {
                 tags = mClassName_TagsMap.get(className);
                 if (tags == null) {
                     tags = new ArrayList<>();
-                    for (Map.Entry<String, BusInfo> entry : mTag_BusInfoMap.entrySet()) {
-                        BusInfo value = entry.getValue();
-                        try {
-                            if (Class.forName(value.className).isAssignableFrom(aClass)) {
-                                tags.add(entry.getKey());
-                                value.classNames.add(className);
+                    for (Map.Entry<String, List<BusInfo>> entry : mTag_BusInfoListMap.entrySet()) {
+                        for (BusInfo busInfo : entry.getValue()) {
+                            try {
+                                if (Class.forName(busInfo.className).isAssignableFrom(aClass)) {
+                                    tags.add(entry.getKey());
+                                    busInfo.classNames.add(className);
+                                }
+                            } catch (ClassNotFoundException e) {
+                                e.printStackTrace();
                             }
-                        } catch (ClassNotFoundException e) {
-                            e.printStackTrace();
                         }
                     }
                     mClassName_TagsMap.put(className, tags);
@@ -158,19 +170,21 @@ public final class BusUtils {
     }
 
     private void postInner(final String tag, final Object arg, final boolean sticky) {
-        BusInfo busInfo = mTag_BusInfoMap.get(tag);
-        if (busInfo == null) {
+        List<BusInfo> busInfoList = mTag_BusInfoListMap.get(tag);
+        if (busInfoList == null) {
             Log.e(TAG, "The bus of tag <" + tag + "> is not exists.");
             return;
         }
-        if (busInfo.method == null) {
-            Method method = getMethodByBusInfo(busInfo);
-            if (method == null) {
-                return;
+        for (BusInfo busInfo : busInfoList) {
+            if (busInfo.method == null) {
+                Method method = getMethodByBusInfo(busInfo);
+                if (method == null) {
+                    return;
+                }
+                busInfo.method = method;
             }
-            busInfo.method = method;
+            invokeMethod(tag, arg, busInfo, sticky);
         }
-        invokeMethod(tag, arg, busInfo, sticky);
     }
 
     private Method getMethodByBusInfo(BusInfo busInfo) {
@@ -273,43 +287,47 @@ public final class BusUtils {
     }
 
     private void postStickyInner(final String tag, final Object arg) {
-        BusInfo busInfo = mTag_BusInfoMap.get(tag);
-        if (busInfo == null) {
+        List<BusInfo> busInfoList = mTag_BusInfoListMap.get(tag);
+        if (busInfoList == null) {
             Log.e(TAG, "The bus of tag <" + tag + "> is not exists.");
             return;
         }
-        if (!busInfo.sticky) { // not sticky bus will post directly.
-            postInner(tag, arg);
-            return;
-        }
-        synchronized (mClassName_Tag_Arg4StickyMap) {
-            Map<String, Object> tagArgMap = mClassName_Tag_Arg4StickyMap.get(busInfo.className);
-            if (tagArgMap == null) {
-                tagArgMap = new HashMap<>();
-                mClassName_Tag_Arg4StickyMap.put(busInfo.className, tagArgMap);
+        for (BusInfo busInfo : busInfoList) {
+            if (!busInfo.sticky) { // not sticky bus will post directly.
+                postInner(tag, arg);
+                return;
             }
-            tagArgMap.put(tag, arg);
+            synchronized (mClassName_Tag_Arg4StickyMap) {
+                Map<String, Object> tagArgMap = mClassName_Tag_Arg4StickyMap.get(busInfo.className);
+                if (tagArgMap == null) {
+                    tagArgMap = new HashMap<>();
+                    mClassName_Tag_Arg4StickyMap.put(busInfo.className, tagArgMap);
+                }
+                tagArgMap.put(tag, arg);
+            }
+            postInner(tag, arg, true);
         }
-        postInner(tag, arg, true);
     }
 
     private void removeStickyInner(final String tag) {
-        BusInfo busInfo = mTag_BusInfoMap.get(tag);
-        if (busInfo == null) {
+        List<BusInfo> busInfoList = mTag_BusInfoListMap.get(tag);
+        if (busInfoList == null) {
             Log.e(TAG, "The bus of tag <" + tag + "> is not exists.");
             return;
         }
-        if (!busInfo.sticky) {
-            Log.e(TAG, "The bus of tag <" + tag + "> is not sticky.");
-            return;
-        }
-        synchronized (mClassName_Tag_Arg4StickyMap) {
-            Map<String, Object> tagArgMap = mClassName_Tag_Arg4StickyMap.get(busInfo.className);
-            if (tagArgMap == null || !tagArgMap.containsKey(tag)) {
-                Log.e(TAG, "The sticky bus of tag <" + tag + "> didn't post.");
+        for (BusInfo busInfo : busInfoList) {
+            if (!busInfo.sticky) {
+                Log.e(TAG, "The bus of tag <" + tag + "> is not sticky.");
                 return;
             }
-            tagArgMap.remove(tag);
+            synchronized (mClassName_Tag_Arg4StickyMap) {
+                Map<String, Object> tagArgMap = mClassName_Tag_Arg4StickyMap.get(busInfo.className);
+                if (tagArgMap == null || !tagArgMap.containsKey(tag)) {
+                    Log.e(TAG, "The sticky bus of tag <" + tag + "> didn't post.");
+                    return;
+                }
+                tagArgMap.remove(tag);
+            }
         }
     }
 
@@ -321,17 +339,19 @@ public final class BusUtils {
         String       paramName;
         boolean      sticky;
         String       threadMode;
+        int          priority;
         Method       method;
         List<String> classNames;
 
         BusInfo(String className, String funName, String paramType, String paramName,
-                boolean sticky, String threadMode) {
+                boolean sticky, String threadMode, int priority) {
             this.className = className;
             this.funName = funName;
             this.paramType = paramType;
             this.paramName = paramName;
             this.sticky = sticky;
             this.threadMode = threadMode;
+            this.priority = priority;
             this.classNames = new CopyOnWriteArrayList<>();
         }
 
@@ -342,6 +362,7 @@ public final class BusUtils {
                     ", sticky: " + sticky +
                     ", threadMode: " + threadMode +
                     ", method: " + method +
+                    ", priority: " + priority +
                     " }";
         }
     }
@@ -358,6 +379,8 @@ public final class BusUtils {
         boolean sticky() default false;
 
         ThreadMode threadMode() default ThreadMode.POSTING;
+
+        int priority() default 0;
     }
 
     private static class LazyHolder {