+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/PhoneUtils.java
@@ -359,28 +359,6 @@ public final class PhoneUtils {
         return false;
     }
 
-    /**
-     * Send sms silently.
-     * <p>Must hold {@code <uses-permission android:name="android.permission.SEND_SMS" />}</p>
-     *
-     * @param phoneNumber The phone number.
-     * @param content     The content.
-     */
-    @RequiresPermission(SEND_SMS)
-    public static void sendSmsSilent(final String phoneNumber, final String content) {
-        if (TextUtils.isEmpty(content)) return;
-        PendingIntent sentIntent = PendingIntent.getBroadcast(Utils.getApp(), 0, new Intent("send"), 0);
-        SmsManager smsManager = SmsManager.getDefault();
-        if (content.length() >= 70) {
-            List<String> ms = smsManager.divideMessage(content);
-            for (String str : ms) {
-                smsManager.sendTextMessage(phoneNumber, null, str, sentIntent, null);
-            }
-        } else {
-            smsManager.sendTextMessage(phoneNumber, null, content, sentIntent, null);
-        }
-    }
-
     private static TelephonyManager getTelephonyManager() {
         return (TelephonyManager) Utils.getApp().getSystemService(Context.TELEPHONY_SERVICE);
     }